+++ lib/src/exceptions/base/port_configuration_exception.dart
@@ -1,0 +1,5 @@
+import 'server_nano_exception.dart';
+
+class PortConfigurationException extends NanoServerException {
+  PortConfigurationException(super.message);
+}