+++ lib/src/utils/logger.dart
@@ -2,8 +2,8 @@ part of '../../server_nano.dart';
 
 void logger(String value, {bool isError = false}) {
   if (isError) {
-    print("Error: $value");
+    developer.log("Error: $value");
   } else {
-    print(value);
+    developer.log(value);
   }
 }